make watcher runnable for all sources
originally we only had kde sources so we had only a need to watch those.
since we aggressively fork and backport nowadays we need much more
comprehensive watching to ensure security updates etc are adopted.

to enable this the watcher gets rejiggered a bit all jobs can now run
watcher but we'll raise (And by extension fail) non-kde software
watching to be dealt with by a developer.

this may still need further adjustment as the merging can conceivably
fall over as well

(also the code is very :@ because it grew over time but was never
refactored to accomodate the grown use cases)