let jobs update filter use a regex
a simple inclusion check is sometimes not all that useful, so allow
using a regex pattern instead.
I'm running the pattern through eval since that kinda asserts validity
of the input, it's also a bit of a crutch though perhaps more ideally
we'd strip the leading and trailing slash and then run the string
through Regexp.new. makes little diff for the end result though