refine version list to pick the none release version
previously we'd ignore versions that aren't released. that had the side
effect that essentially products were skipping validation in general
when they had a pending release.
this is silly though. up until the release is public we expect the
previous version to be in the archive.

this still has some beauty problem for snapshotting because whether or
not the latest kf5 is in the snapshot is immaterial to the quality of
the snapshot, so... there's that