add new tech to auto-generate appstream filters
this is currently not wired into any job or anything

the motivation here is that we keep having users lament how foobar
doesn't install when foobar is a) not kde software so we don't care
anyway and b) foobar is possibly using qt private ABI breaking compat

it makes for a shitty experience that we have a hard time dealing with
short of selectively mirroring ubuntu, which seems an awkward solution
as well as we'd lose part of the appeal of sitting on top of the giant
that is ubuntu lts.

sooooo as a somewhat immediate measure to improve visual user experience
we'll want to exclude all applications coming from ubuntu from the
software stores.

this has a number of effects.
a) if the software is avialable in a bundle format the user would get
that instead hopefully leading to greater adoption and in turn greater
interest in fixing pain points if there are any
b) the problem outlined above goes away and ideally the user will now
only find software that actually works
c) kde software that changes id won't show up twice (e.g. ubuntu has
org.kde.krita.desktop but then it was renamed to org.kde.krtia and the
devs neglected to set the legacy name hint so appstream doesn't know
that they the same thing - since we'll filter the ubuntu one out there
can only be one krita and that is ours)

as such we filter addons and desktop applications now (addons may
actually be illadvised as I guess you could addon into any other
component, in practise I believe this will only be applications though,
so I doubt it matters much)

the way this is going to work in production is that a job will clone the
metainfo repo, run the scripty on it, the scripty injects the detected
components into the removal list, job pushes delta into repo. the script
also keeps a pristine list of auto removed components so it can on the
next run tell which components are there because a human added them and
which ones were auto-added. I'm hoping the last part will be sufficient
to prevent awkward clashing between human and computer as otherwise
we'll need some sort of intermediate file structure with distinct files
that gets compiled into the final removal file before running asgen