fix repo
/dev/unstable now is a symlink to /unstable but we need the origin path.
the path here is not only used to download stuff but also to construct
the skel path where asgen push needs to push to. the problem there being
that since that was pointing to the symlink location the generation
itself (the download) would work but the publishing would write to the
wrong skel which meant that the appstream metadata doesn't actually
update.

this was never correctly ported to the non-/dev/ paths -.-