force invent urls to use .git
we have rate limiting on requests towards invent so we shouldn't spend
needless requests asking for urls that gitlab will then have to redirect
to $url.git anyway. this is a bit hacky in that urls have multiple
sources that will inconsistently use the suffix. to cover everything the
SCM class itself will actually force the suffix in which has the
disadvantage that it is inobvious mutation but at the same time the
advantage of being innobvious mutation meaning other components don't
need to worry about whether the url is awesome or not.