let core defaults go to 8 for nci
this addresses a number of problems to do with lack of data.

firstly https://invent.kde.org/teams/neon/-/issues/43 which is actually
also the generic reason why we want to start at the top of the core
profiles.

a new job (such as future_series jobs) has no metrics so it'd start off
at 2core but that means the job will have horrible performance for N
builds (currently N>=2) before it gets upgraded to the next core profile
and again have horrible performance for N builds before it gets upgraded
again... etc etc

this is particularly also is a problem for the qt experimental-staging
builds because they may get thrown away when not needed and created
again when needed meaning they aren't in the scoring json files but
instead get their default assignment from binarier and thus bypass the
override rules in the scoring system (the job hasn't been scored because
the job doesn't exist)

on top of that new jobs will be stuck on bad performance for however
long it takes to run the job updater again. that may be literal months
of bad performance.

by defaulting to 8 we'll have fast performance right out the gate but
downgrade the core profile on subsequent runs