add experimental tech to bypass jenkins api
the api is very slow, when run on the master we could possibly gain a
lot of speed by bypassing the api and poking the files instead.

since technically only ever the job updater should update jobs there's
no huge concern over reentrantcy here to the point where we could even
run updates through the local disk and then only tell jenkins to reload
itself once we are done (although reloads are also somewhat slow so
there's that)

for starters lets read the config through the local file. this should
actually be fairly cool already because noop updates (which are most,
most of the time) will not talk to the api at all then