move job scoring out of job updater
it's actually not nearly as fast as one would wish and so we want it run
randomly during the week. whenever job updater runs it will have
reasonably up to date data so core changes should still happen, albeit
not nearly as timely. but that's the cost of faster job updating I guess