try to remove colord from iso
I'm actually not a 100% confident this is going to work but from looking
at the logs the package list is installed in a single apt-get call.

as such we should be able to mark a package from the set for removal
(i.e. not getting it dragged in at all)