rejigger torrent support
the mirror software serving files.kde.org changed and the new thingy
doesn't support metalinks. it does however send a limited number of
mirrors in the http headers of the file so we can grab them from there.
at the same time the python tool now makes even less sense so import its
functionality into the main script. arguably I'd like this more if it
was a class now and unit tested but meh no time