do not deploy tooling on shared nodes
e.g. a server with both armhf and arm64

tooling deploys to $HOME and manages gems in the home so it must only
run on one of the nodes.

for that purpose the label shared-node is used to tell when a node is
shared on another node (e.g. armhf-01 is same host as arm64-01 so the
former has shared-node and the latter doesn't effectively running
tooling deploy only once)