do not delete running containers younger than a day
this actually was racing.

mgmt_docker could sweep in and delete a container between another job
having created it and trying to clean up (if the container was
ephemeral, which of course all our containers are). the container was
exited but still necessary.

ultimately I suppose we can simply make the cleanup run only after a
day. a container that is running after a day definitely shouldn't be
around anyway.