constraint gem versions to make them ruby 2.5 compatible
jenkins_api_client depends mixlib-shellout depends chef-utils depends
ruby 2.6.

we actually don't need this as it's only used for the jar cli app which
we don't use. so even if this broke we'd not care much, but since it's
marked as a dep in the api client we need to force it onto compatible
versions