install debhelper when setting up repos
since we have the call for pkg-kde-tools anyway we can use it to pull in
debhelper

this can speed up sourcing. the way sourcing works is that it first
tries to create the source without build deps installed at the off
chance that the bare minimum for cleaning the source is already
available in the base image. this fails aaaalll the time for many
sources that use dh without pkg-kde-tools. yet it's very likely all that
was missing is dh itself.

so lets install dh always to improve the chance of getting a speedier
sourcing