remove version override code
there are a number of problems with this but I'm removing it because
Bhushan asks for it because the number of requests are a problem for
sysadmins and it's not getting fixed by anyone apparently

further wrong

- hardcoding of series names and paths
- downloading assets through webuis instead of git (mind you, git would
be wrong too, albeit differently wrong)
- unnamed regex group access
- using class variables instead of class-instance variables
- calling cached getters and then using the underlying (internal)
variables like they aren't getters
- not checking versions at all vis a vis the pins even though the very
purpose of these checks are to verify versions
- not handling http error codes
- not actually testing the pins. what likely should happen is the checks
installing neon-settings and friends and then handling this via
apt-cache. i.e. the actual check constraint shouldn't be skipped but
still met! (the apt-cache should candidate a suitable version)