excessively refactor the versions linting
still far from where I want or need it to be, but a start

this piece of tech evolved rather unfortunately. I started off with all
sorts of stuff in the same file which nobody ever fixed so it was a huge
pile of stuff that shouldn't inter depend nearly as much

so first and foremost the individual classes are now in their own files
and consequently the listers have their own standalone tests now. this
then allows us to have much higher level mocking the versions test since
we aren't concerned with testing the listers there

this also adds a new pin lister but it's not used anywhere as I am yet
unsure how or where it'd be used. after a lot of soul searching I'm
leaning towards arguing that we should never want to know what the pins
are. for distro upgrades 20.04->22.04 it's a none concern as we only
care newversion>>oldversion we can simply install the pins for that and
look at the candidates. for the post-build testing we expect
newbuild>>archive which is also unaffected by pins since a new build
that isn't large than what is in the archive is always a problem on some
level. lastly the mass tests allourversions>>ubuntu are the only ones
that actually have use for pins but it's hugely complicated because we
*do* care if **our** exact version is actually the one that is candidate
so we'd have to apply a whole bunch of otherwise useless constraints
just to deal with pins that are making a lower version from us outscore
a higher version from ubuntu. it's needless complexity and I think we'll
all be happier if we simply stick to allourversions>>ubuntu there are
numerous ways of achieving it anyway

moving on, the tests are now split in different files which means we
don't have two test classes inside the same scope. when testing versions
we test versions, not distro upgrade version, and vice versa.

also the way the of version test is started now is different. the
lister= has become an init and it gets fed both the ours and theirs
packages list respectively obtained from the lister implementations.
this simplifies test mocking and also makes the way we obtain and handle
listings consistent (previously theirs listing was obtained inside
versioncheck through a class level cache - that also meant things were
slower than they needed to be because we'd always have to iter the
entire cache N times; N=number of tests)