simplify mapping function
we always mangle errors into warnings by default so we don't need the
hack for source-is-missing anymore. the plan here kind of is that
nothing can be an error unless we explicitly promote it to error. this
is because lintian is overzealous with tagging things as errors to the
point where all errors it might emit on a package is of no interest to
us. so I ended up turning everything into warnings by default and we
have to opt into fatality for certain tags