move lintian from log linter to bin linter
ever since the original lint tech was written we've relied on
pkg-kde-tools to run lintian for us, this was fine in 2016 when we only
built KDE software and had no notion of ever changing that. requirements
have evolved a lot since then so we now build lots of software which is
not actually made by KDE and by extension not using pkg-kde-tools. this
left us with lintian not running on third party software which we still
have because it is somehow playing a part in our software so we *do*
care about its quality.

to that end we are now running lintian ourselves instead of relying on
pkg-kde-tools to do it for us. also install a binary overlay to noop
part of the auto-lintian runs. unfortunately there are two ways it can
get invoked but one isn't easily disabled :(

this also sees the move of the linter from log linting to binary linting
(functional difference is that one parses the log out put the and the
other does direct magic on the file tree). as a result the linter now
manually invokes genchanges and lintian (exit code sensitivity might
still need adjustments for genchanges, currently I'm thinking it should
always pass but we'll see).

testing is largely the same except there is now no invalidity scenario.
lintian always runs and we always know what its output is, if the output
contains nothing of value that's still a valid result

also the linter is not run through the blackbox lint_bin test because
mocking the lintian shebang is fairly impossible right now, well,
certainly hard and not worth the effort anyway.