turn distinctly bad sonames into errors
we've had it happen that the soname libkColorPicker0.1.4 was packaged as
libkcolorpicker0 rather than libkcolorpicker0-1-4.
this then lead to breakage when an update rolled around and the soname
of the library changed but the binary package wasn't transitioned
accordingly because seemingly everything was passing as far as the
snapshot QA was concerned.

this also implicitly helps guard against cases where the install file is
inappropriately aggressive with wildcarding as a library package must
really never be installing usr/lib/*/lib*.so.* it falls on our heads
since we don't have extra ABI tracking and solely rely on upstream
so-versioning correctly but that still requires us to learn of changes
to the so-version which is why the wildcarding is absolutely never
appropriate for us. there'll probably a more aggressive test for this in
due time but until then this implicit check is helping us already