revise lintian tech yet more
the previous plan was foiled by the fact that we split apart the binary
and source and since there's no real good way to deal with this I've
decided to instead embrace the difference.

there is now a binaries test that exclusively runs on the result data of
the binary job. this means we don't need to generate a changes file and
can simply use the one that was actually generated by the build.

the downside is that we lose linting of the sources for now. at the same
time that's also a bit of an advantage though. previously we'd have
linted the source for every arch, but since we have split changes files
we can also do split linting i.e. we could lint the source in the _src
job now.

in conclusion source linting will need further work still