port away from ci-tooling
this has been deprecated since 2014 and nothing was done about it so.
here I am jumping in the deep end.

this may well blow up since even newer programs have been using
ci-tooling or worse lived in ci-tooling while using libs from the main
dir. so we basically have to rely on test coverage for this but that too
doesn't exist for many of the programs, libraries meanwhile are fairly
ok so I have no worries there.

on the plus side fixing regressions should be easy enough as it simply
means changing the relative reference so it is correct for the new
structure and that is trivial because everything is where it was, the
ci-tooling/ subdir has simply been folded up.

to not have to regenerate all jenkins jobs we still have compatibility
symlinks for the programs. libraries are intentionally excluded because
they can cause require loops between a symlink library and its real
file, also double requires - ruby doesn't consider the symlink and the
real file the same path so they always get required even when require
usually would prevent double-loading the same file