change build type api to abstract type decision
this allows us to more generally do a check on whether release build has
been switched on or not. specifically since we strip flags from dpkg
when doing release builds the ddebs would be empty and cause a mountain
of lintian warnings. to avoid that we can simply skip ddeb generation as
a whole