add a config hook to assert livecd-rootfs is from neon
> unknown project neon

it has been a great source of confusion over the years what exactly is
wrong when that error appears and it's an entirely daft reason: our
livecd-rootfs has fallen behind ubuntu's and needs a merge

we already had a forks assertion as a build hook the trouble is that
build hooks run sometime mid-build while livecd-rootfs needs to be
correct right from the get go. fortunately config hooks run even before
live-build and so should be able to preemept livecd-rootfs being from
the wrong origin

for simplicities sake the test is based on the fact that the Maintainer
is always neon for neon builds. establishing the origin through apt is a
bit fiddly because none of the outputs are easy to grep and 'show'
specifically will show the candidate not the installed version so
there's opportunity for racing there.