loosen mocha dep
when mocha breaks it's not the biggest deal since it's only used in
tests plus we'll know because of just that case

this update also revealed that we freeze incoming objects in the project
job. this is a bit fishy, dup them first. I don't quite recall why we
freeze but I guess it doesn't hurt