update the docker related VCR casettes
this also required a slight adjustment to the containment failure test.
I am not super sure that it won't cause adverse affects but given
failure is failure I doubt it.

this update isn't technically necessary but since the previous commit
integrates a huge docker gem update I feel like it makes sense to rerun
the test set from scratch without recordings to ensure there's nothing
wonky going on (like that new exception for example ;)). so here are the
new casettes though they are largely the same as the old ones