start ripping apart the watcher
this one long function was aaaaaaaaaaaaaaaabsolutely horrible. so now
it's divided up into multiple methods.

this is still very horrible because there are many methods that set
arbitrary instance variables on the watcher or even mutate them in
place. omg.

there now is an assertion being made that there is no more than one dehs
package. in the past we had multiple lines per watch file resulting in
multiple packages getting returned (notably this was used for stable and
unstable urls of KDE software). this since has been removed in KDE and I
have high hopes that no other source uses this crap either so we can
greatly simplify things by not having to worry about filtering multiple
DEHS packagse along the way.

other than this is largely the same as the previous code, albeit split
up.

notification mails have had their subject line fixed up to not be so
darn similar and so we know what needs doing

also encountering unstable urls now raises instead of returns as
discussed on IRC. personally I'd do away with the check but Jon says it
still makes sense to prevent unstable KDE software from entering user
edition (arguably this should be happening in the actual build job, not
the watcher, but in lieu of moving the code there we'll simply raise
here for now)

changelog.rb has also seen some reshaking because we need access to the
raw command line string so that we can give developers a one-liner for
how to integrate the new release. kinda meh but whatever

all system calls have been changed to run through tty command which has
the advantage that failure of the git commands will actually show up as
failing jobs (instead of getting silently ignored)