add build type to version string
there's a bug in aptcc that prevents it from accurately detecting that
v5 is different from v5 when they were different builds but had the same
version. as can happen with our previous versions because different
editions would construct the same build versions. now they are different
and should allow aptcc to detect them as different