kick out the remote naming
I am not sure why this is there, it's very legacy though and I've traced
it back to KCI times. I think :tm: this used to be used before we
switched to multi scm. we had both the packaging and the source in the
same repo but as different remotes (one packaging, one upstream). if
that is indeed why it is there it's not been relevant in years.

I've removed it temporarily already and seen no adverse effect in the
few builds I've kicked, so I'm guessing it's safe to roll out