skip qa on all of experimental
since we stage qt in a special type that isn't used for anything else we
can skip over it all wholesale. this specifically disables the
linqml/cmake jobs for the non-qt builds that we have inside experimental
(kwin for example). since experimental isn't a fully functionally type
the package-dependent linters are always red there, turning the entire
job red, making it unnecessarily hard to see if they build (which is the
main concern for experimental staging)