bump ulimit and hijack the binary
a) qtwebengine calls ulimit unconditionally
b) qtwebengine fails to build cause ulimit may not be permitted in all
our dockers
...
c) overlay ulimit with /bin/true
d) bump our overall limit to 4096 (what QWE has)

should this not be enough in the future we can always revisit this
approach. it is a bit hacky, but then I am fairly sick of QWE weirdness
and conceptually speaking we any package calling ulimit can sod right
off, so /bin/true it is.