pull node name into container name for xci
xci uses pipeline builds that share the same job_name for all
architectures. this is a problem because arm64 and armhf are on the same
server, so one would clobber the other's work because they both want to
use the same container name. by injecting the node name the container
name should be unique (and persistent) once more