refactor cmake linting
this broke again -.-

instead of using segmentation add an extractor hack to the cmakelists.
this is a bit less broad in scope because we'll only see packages that
are  actually reported by feature_summary (previously we'd detect
everything spewed into the output so long as it matched one of the
expressions) but over all should proof much more reliable then the
segmentation that breaks every time the marker lines get changed/removed
or the sequence as a whole changes. it's supremely annoying that we need
so much code but whatevs.

this new approach sports a FeatureSummaryExtractor inspired by
KCrashLinkValidator. it appends a feature_summary call to the cmakelists
that writes the output to file rather than stdout. this enables us to
get precise access to the summary in individual files.

this also required substantial rejiggering on the lint side. the cmake
linting is now part of packaging linting not log linting. the extractor
is installed in the shared package builder which I hope won't blow up
the neon CIs

there may actually be some problems with this commit, it's exceptionally
hard to synthesize tests for this