start always including arch tags
this is all a bit awkward because architecture permutations were only
added to the imaging tech after the fact. revise the way this works with
the new dockerhub based workflow. always tag all images with their
architecture as well so we can consistently fetch the right
architectures and use the right architectures (in particular of note for
arm64 where we also have armhf images)